fix test running
