+++ src/ShapeCrawler/Charts/IAxis.cs
@@ -1,32 +1,34 @@
 ﻿using System;
+using System.Linq;
 
-// ReSharper disable once CheckNamespace
 namespace ShapeCrawler;
 
+using C = DocumentFormat.OpenXml.Drawing.Charts;
+
 /// <summary>
-///     Represents axis bounds.
+///     Represents a chart axis.
 /// </summary>
-public interface IBounds
+public interface IAxis
 {
     /// <summary>
-    ///     Gets or sets the minimum value of the axis.
+    ///     Gets or sets axis minimum value.
     /// </summary>
     double Minimum { get; set; }
-
+    
     /// <summary>
-    ///     Gets or sets the maximum value of the axis.
+    ///     Gets or sets axis maximum value.
     /// </summary>
     double Maximum { get; set; }
 }
 
-internal class SCBounds : IBounds
+internal class SCAxis : IAxis
 {
     private const double DefaultMax = 6;
-    private readonly DocumentFormat.OpenXml.Drawing.Charts.Scaling cScaling;
+    private readonly C.PlotArea cPlotArea;
 
-    public SCBounds(DocumentFormat.OpenXml.Drawing.Charts.Scaling cScaling)
+    public SCAxis(C.PlotArea cPlotArea)
     {
-        this.cScaling = cScaling;
+        this.cPlotArea = cPlotArea;
     }
 
     public double Minimum
@@ -46,12 +48,6 @@ internal class SCBounds : IBounds
         throw new NotImplementedException();
     }
 
-    private double GetMaximum()
-    {
-        var cMax = this.cScaling.MaxAxisValue;
-        return cMax == null ? DefaultMax : cMax.Val!;
-    }
-
     private void SetMinimum(double value)
     {
         throw new NotImplementedException();
@@ -59,7 +55,17 @@ internal class SCBounds : IBounds
 
     private double GetMinimum()
     {
-        var cMin = this.cScaling.MinAxisValue;
+        var cScaling = this.cPlotArea.Descendants<C.Scaling>().First();
+        var cMin = cScaling.MinAxisValue;
+        
         return cMin == null ? 0 : cMin.Val!;
+    }
+    
+    private double GetMaximum()
+    {
+        var cScaling = this.cPlotArea.Descendants<C.Scaling>().First();
+        var cMax = cScaling.MaxAxisValue;
+        
+        return cMax == null ? DefaultMax : cMax.Val!;
     }
 }
