+++ src/ShapeCrawler/Charts/SCChart.cs
@@ -107,7 +107,6 @@ internal class SCChart : SCShape, IChart
 
     public SpreadsheetDocument SDKSpreadsheetDocument => this.ChartWorkbook!.SpreadsheetDocument.Value;
     
-    public IFormatAxis FormatAxis => this.GetFormatAxis();
     public IAxesManager Axes => this.GetAxes();
 
     internal ChartWorkbook? ChartWorkbook { get; set; }
@@ -140,11 +139,6 @@ internal class SCChart : SCShape, IChart
         Enum.TryParse(chartName, true, out SCChartType enumChartType);
 
         return enumChartType;
-    }
-    
-    private IFormatAxis GetFormatAxis()
-    {
-        return new SCFormatAxis(this.cPlotArea);
     }
     
     private IAxesManager GetAxes()
@@ -233,66 +227,4 @@ internal class SCChart : SCShape, IChart
         return this.cXCharts.First().ChildElements
             .FirstOrDefault(e => e.LocalName.Equals("ser", StringComparison.Ordinal));
     }
-}
-
-internal class SCAxesManager : IAxesManager
-{
-    private readonly C.PlotArea cPlotArea;
-
-    public SCAxesManager(C.PlotArea cPlotArea)
-    {
-        this.cPlotArea = cPlotArea;
-    }
-
-    public IAxis HorizontalAxis => new SCAxis(this.cPlotArea);
-}
-
-internal class SCAxis : IAxis
-{
-    private const double DefaultMax = 6;
-    private readonly C.PlotArea cPlotArea;
-
-    public SCAxis(C.PlotArea cPlotArea)
-    {
-        this.cPlotArea = cPlotArea;
-    }
-
-    public double Minimum
-    {
-        get => this.GetMinimum();
-        set => this.SetMinimum(value);
-    }
-
-    public double Maximum
-    {
-        get => this.GetMaximum();
-        set => this.SetMaximum(value);
-    }
-
-    private void SetMaximum(double value)
-    {
-        throw new NotImplementedException();
-    }
-
-    private void SetMinimum(double value)
-    {
-        throw new NotImplementedException();
-    }
-
-    private double GetMinimum()
-    {
-        var cScaling = this.cPlotArea.Descendants<C.Scaling>().First();
-        var cMin = cScaling.MinAxisValue;
-        
-        return cMin == null ? 0 : cMin.Val!;
-    }
-    
-    private double GetMaximum()
-    {
-        var cScaling = this.cPlotArea.Descendants<C.Scaling>().First();
-        var cMax = cScaling.MaxAxisValue;
-        
-        return cMax == null ? DefaultMax : cMax.Val!;
-    }
-
 }
